Move deferred value classes to correct package
