+++ src/main/java/com/hubspot/jinjava/interpret/DeferredValue.java
@@ -1,0 +1,4 @@
+package com.hubspot.jinjava.interpret;
+
+public class DeferredValue {
+}
