--- src/main/java/com/hubspot/jinjava/interpret/errorcategory/DeferredValue.java
@@ -1,4 +1,0 @@
-package com.hubspot.jinjava.interpret.errorcategory;
-
-public class DeferredValue {
-}
